Add shape tests for render and shallow apis
